Updated passHistory with Bootstrap table
The table portion of the html is complete, but this file still needs the design to be reworked. The JavaScript for this file is already able to get a JSON string from the server and is also able to update the table given a js array, and is only missing the step in between where the JSON string is parsed. This part of the app is near finished and should be done by Monday when I can access the server.